+++ firmware/main/app_config.h
@@ -87,6 +87,8 @@
 #include "boards/waveshare_epaper_397.h"
 #elif defined(CONFIG_DRAFTLING_MODEL_SEEED_RETERMINAL_E1001)
 #include "boards/seeed_reterminal_e1001.h"
+#elif defined(CONFIG_DRAFTLING_MODEL_SEEED_RETERMINAL_STICKY)
+#include "boards/seeed_reterminal_sticky.h"
 #elif defined(CONFIG_DRAFTLING_MODEL_M5STACK_PAPERMONO)
 #include "boards/m5stack_papermono.h"
 #else